Rendering now works!
Apparently there was a bogus check in modelread.cpp. This has been
removed. The old way of doing thruster anims should still work, but to
take advantage of multiple thruster anims, the anims should be placed in
maps instead of effects and the "use thruster materials" flag should be
added to the ship entry.